data_handler/requirements.txt
confluent_kafka==2.3.0
flask==2.1.4
sqlite3